Restructuring the Folder Structure
