+++ src/App.css
@@ -6,256 +6,15 @@
 .download-button {
     margin: .5rem;
 }
-/* HOMEPAGE */
-
-#home {
-    display: flex;
-    flex: 1;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-	margin-bottom: 4rem;
-}
-
-#home #version-text {
-    color: hsl(0, 0%, 45%);
-    font-size: .7rem;
-    margin-bottom: .1rem;
-}
-
-#home a {
-    text-decoration: none;
-}
-#home h1 {
-    font-size: 3.6rem;
-    margin-bottom: .2rem;
-    color: hsl(0, 0%, 10%)
-}
-#home h1:hover {
-    color: rgb(26, 153, 130);
-}
-
-#home p {
-    font-size: 1.4rem;
-    margin-bottom: 1.4rem;
-    text-align: center;
-    color: hsl(0, 0%, 20%);
-}
-
-#home #cta-buttons {
-    display: flex;
-    flex-direction: row;
-}
-
-#home #cta-buttons #main-button {
-    background-color: hsl(219, 86%, 57%);
-    padding: .6rem 1.6rem;
-    border-radius: .5rem;
-    color: hsl(0, 0%, 100%);
-    text-decoration: none;
-    font-size: 1.2rem;
-    margin-right: .5rem;
-    transition: box-shadow 250ms;
-}
-#home #cta-buttons #main-button:hover {
-    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
-}
-
-#home #cta-buttons #gh-button {
-    display: block;
-    border: 1px solid hsl(0, 0%, 69%);
-    padding: .6rem .6rem .45rem;
-    border-radius: .5rem;
-    color: hsl(0, 0%, 100%);
-    text-decoration: none;
-    transition: box-shadow 250ms, border 100ms;
-}
-#home #cta-buttons #gh-button:hover {
-    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
-    border: 1px solid transparent;
-}
-
 
 /* COMPONENTS */
-
 #root {
     display: flex;
     height: 100vh;
     justify-content: space-between;
 }
-/* NAVIGATION */
-
-#navigation {
-    padding: .8rem;
-    margin: auto 1rem;
-    border-top-right-radius: .5rem;
-    border-bottom-right-radius: .5rem;
-    border-top: 1px solid hsl(0, 0%, 78%);
-    border-right: 1px solid hsl(0, 0%, 78%);
-    border-bottom: 1px solid hsl(0, 0%, 78%);
-    background-color: hsl(0, 100%, 100%);
-}
-
-#navigation #separator {
-    border: 1px solid hsl(0, 0%, 90%);
-}
-
-#navigation ul {
-    list-style-type: none;
-}
-
-#navigation ul li {
-    width: 10rem;
-    margin-bottom: .4rem;
-}
-#navigation ul li:last-child {
-    margin-bottom: 0;
-}
-
-#navigation ul li a {
-    display: block;
-    font-size: 1.025rem;
-    color: hsl(0, 0%, 40%);
-    background-color: transparent;
-    padding: .6rem;
-    border-radius: .5rem;
-    text-decoration: none;
-    transition: color, background-color 250ms;
-}
-
-.style-remover {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-.style-remover .link {
-    text-decoration: none;
-}
-.list-inline {
-    overflow: hidden;
-    border-radius: .5em;
-}
-.list-inline li {
-    float: left;
-}
-
-#home-button {
-    font-weight: 700;
-    font-size: 1.5em;
-    text-align: center;
-    padding: 0;
-    margin: 0;
-    display: block;
-    padding: 1em 0;
-    color: hsl(0, 0%, 0%);
-    text-decoration: none;
-    transition: 0.15s linear color;
-}
-#home-button:hover {
-    color: hsl(169, 71%, 35%);
-}
-
-#result-button {
-    background-color: hsl(219, 86%, 57%);
-    font-weight: 700;
-    font-size: 1.5em;
-    text-align: center;
-    margin-top: 1em;
-    display: block;
-    padding: .5em 0;
-    border-radius: .5rem;
-    color: hsl(0, 0%, 100%);
-    text-decoration: none;
-    font-size: 1.2rem;
-    margin-right: .5rem;
-    transition: box-shadow 250ms;
-}
-#result-button:hover {
-    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
-}
-
-.list-hover-slide li {
-    position: relative;
-    overflow: hidden;
-    border-radius: .5em;
-}
-
-.list-hover-slide .link {
-    display: block;
-    position: relative;
-    z-index: 1;
-    transition: 0.35s ease color;
-}
-
-#navigation .link:hover{
-    color: hsl(0, 0%, 100%);
-}
-
-.list-hover-slide .link:before {
-    content: '';
-    display: block;
-    z-index: -1;
-    position: absolute;
-    left: -100%;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    border-right: solid 5px hsl(0, 0%, 0%);
-    background: hsl(219, 86%, 57%);;
-    transition: 0.35s ease left;
-}
-
-.list-hover-slide .link.is-current:before, .list-hover-slide .link:hover:before {
-    left: 0;
-}
-
-/* PAGES */
-
-/* PAGE NOT FOUND */
-
-#error-container{
-    position : fixed;
-    top : 50%;
-    left : 50%;
-    transform: translate(-50% , -50%);
-    text-align: center;
-    padding : 1rem;
-    width : 60%;
-}
-
-#error-container #error-status-code {
-    font-size : 72px;
-}
-
-#error-container #error-title{
-    margin-bottom : 10px;
-    font-size : 24px;
-}
-
-#error-container #error-content{
-    font-size : 18px;
-    margin-bottom : 10px;
-}
-
-#error-container #home-button{
-    background-color: hsl(219, 86%, 57%);
-    padding: .6rem 1.6rem;
-    border-radius: .5rem;
-    color: hsl(0, 0%, 100%);
-    text-decoration: none;
-    font-size: 1.2rem;
-    transition: box-shadow 250ms;
-    display : inline-block;
-}
-
-#error-container #home-button:hover{
-    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
-    border: 1px solid transparent;
-}
-
 
 /* SECTION */
-
 .section-wrapper {
     display: flex;
     flex: 1;
@@ -398,62 +157,5 @@
     color: hsl(219, 86%, 57%);
 }
 #preview-text span a:hover {
-    text-decoration: underline;
-}
-
-/* PRIVACY POLICY */
-
-#privacy-policy {
-    margin: 2rem auto;
-    width: 50%;
-}
-
-#privacy-policy a {
-    text-decoration: none;
-    color: hsl(219, 86%, 57%);
-}
-#privacy-policy a:hover {
-    text-decoration: underline;
-}
-
-#privacy-policy > h1, h2 {
-    margin-bottom: .8rem;
-    margin-top: .4rem;
-}
-
-#privacy-policy #footer {
-    position: initial;
-	margin-top: 2rem;
-}
-/* RESULT */
-#result {
-    display: flex;
-    align-items: center;
-}
-/* FOOTER */
-
-#footer {
-    position: fixed;
-    bottom: 0;
-}
-
-#footer > ul {
-    margin-bottom: 1.5rem;
-    list-style-type: none;
-    display: flex;
-    flex-direction: row;
-}
-#footer > ul li {
-    margin-right: .7rem;
-}
-#footer > ul li:last-child {
-    margin-right: 0;
-}
-
-#footer > ul li a {
-    text-decoration: none;
-    color: hsl(219, 86%, 57%);
-}
-#footer > ul li a:hover {
     text-decoration: underline;
 }
