+++ src/components/footer/Footer.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
+import './Footer.css';
+
 function Footer() {
     return (
         <div id="footer">
