+++ src/components/navigation/Navigation.css
@@ -1,0 +1,123 @@
+#navigation {
+    padding: .8rem;
+    margin: auto 1rem;
+    border-top-right-radius: .5rem;
+    border-bottom-right-radius: .5rem;
+    border-top: 1px solid hsl(0, 0%, 78%);
+    border-right: 1px solid hsl(0, 0%, 78%);
+    border-bottom: 1px solid hsl(0, 0%, 78%);
+    background-color: hsl(0, 100%, 100%);
+}
+
+#navigation #separator {
+    border: 1px solid hsl(0, 0%, 90%);
+}
+
+#navigation ul {
+    list-style-type: none;
+}
+
+#navigation ul li {
+    width: 10rem;
+    margin-bottom: .4rem;
+}
+#navigation ul li:last-child {
+    margin-bottom: 0;
+}
+
+#navigation ul li a {
+    display: block;
+    font-size: 1.025rem;
+    color: hsl(0, 0%, 40%);
+    background-color: transparent;
+    padding: .6rem;
+    border-radius: .5rem;
+    text-decoration: none;
+    transition: color, background-color 250ms;
+}
+
+.style-remover {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+.style-remover .link {
+    text-decoration: none;
+}
+.list-inline {
+    overflow: hidden;
+    border-radius: .5em;
+}
+.list-inline li {
+    float: left;
+}
+
+#home-button {
+    font-weight: 700;
+    font-size: 1.5em;
+    text-align: center;
+    padding: 0;
+    margin: 0;
+    display: block;
+    padding: 1em 0;
+    color: hsl(0, 0%, 0%);
+    text-decoration: none;
+    transition: 0.15s linear color;
+}
+#home-button:hover {
+    color: hsl(169, 71%, 35%);
+}
+
+#result-button {
+    background-color: hsl(219, 86%, 57%);
+    font-weight: 700;
+    font-size: 1.5em;
+    text-align: center;
+    margin-top: 1em;
+    display: block;
+    padding: .5em 0;
+    border-radius: .5rem;
+    color: hsl(0, 0%, 100%);
+    text-decoration: none;
+    font-size: 1.2rem;
+    margin-right: .5rem;
+    transition: box-shadow 250ms;
+}
+#result-button:hover {
+    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
+}
+
+.list-hover-slide li {
+    position: relative;
+    overflow: hidden;
+    border-radius: .5em;
+}
+
+.list-hover-slide .link {
+    display: block;
+    position: relative;
+    z-index: 1;
+    transition: 0.35s ease color;
+}
+
+#navigation .link:hover{
+    color: hsl(0, 0%, 100%);
+}
+
+.list-hover-slide .link:before {
+    content: '';
+    display: block;
+    z-index: -1;
+    position: absolute;
+    left: -100%;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    border-right: solid 5px hsl(0, 0%, 0%);
+    background: hsl(219, 86%, 57%);;
+    transition: 0.35s ease left;
+}
+
+.list-hover-slide .link.is-current:before, .list-hover-slide .link:hover:before {
+    left: 0;
+}
