+++ src/components/navigation/Navigation.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
+import "./Navigation.css";
+
 function Navigation() {
     return (
