+++ src/pages/education/Education.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import Navigation from '../components/Navigation';
-import useLocalStorage from '../scripts/localStorageData';
+import Navigation from '../../components/navigation/Navigation';
+import useLocalStorage from '../../scripts/localStorageData';
+import "./Education.css";
 
 function Education()  //Defining the function Education in which all the constant defined which uses to store the values of User's Education Information 
 {
