+++ src/pages/home/Home.css
@@ -1,0 +1,66 @@
+#home {
+    display: flex;
+    flex: 1;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+	margin-bottom: 4rem;
+}
+
+#home #version-text {
+    color: hsl(0, 0%, 45%);
+    font-size: .7rem;
+    margin-bottom: .1rem;
+}
+
+#home a {
+    text-decoration: none;
+}
+#home h1 {
+    font-size: 3.6rem;
+    margin-bottom: .2rem;
+    color: hsl(0, 0%, 10%)
+}
+#home h1:hover {
+    color: rgb(26, 153, 130);
+}
+
+#home p {
+    font-size: 1.4rem;
+    margin-bottom: 1.4rem;
+    text-align: center;
+    color: hsl(0, 0%, 20%);
+}
+
+#home #cta-buttons {
+    display: flex;
+    flex-direction: row;
+}
+
+#home #cta-buttons #main-button {
+    background-color: hsl(219, 86%, 57%);
+    padding: .6rem 1.6rem;
+    border-radius: .5rem;
+    color: hsl(0, 0%, 100%);
+    text-decoration: none;
+    font-size: 1.2rem;
+    margin-right: .5rem;
+    transition: box-shadow 250ms;
+}
+#home #cta-buttons #main-button:hover {
+    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
+}
+
+#home #cta-buttons #gh-button {
+    display: block;
+    border: 1px solid hsl(0, 0%, 69%);
+    padding: .6rem .6rem .45rem;
+    border-radius: .5rem;
+    color: hsl(0, 0%, 100%);
+    text-decoration: none;
+    transition: box-shadow 250ms, border 100ms;
+}
+#home #cta-buttons #gh-button:hover {
+    box-shadow: 0px 5px 8px 3px rgba(0,0,0,0.2);
+    border: 1px solid transparent;
+}
