+++ src/pages/home/Home.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Footer from '../components/Footer';
+import Footer from '../../components/footer/Footer';
+import "./Home.css";
 
 function Home() {
     return (
