+++ src/pages/job/Job.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import Navigation from '../components/Navigation';
-import useLocalStorage from '../scripts/localStorageData';
+import Navigation from '../../components/navigation/Navigation';
+import useLocalStorage from '../../scripts/localStorageData';
+import "./Job.css";
 
 function Job() {                       //Defining the function Job in which all the constant defined which uses to store the values of User's Jobs Information 
     const [valueJobCompany1, setValueJobCompany1] = useLocalStorage(
