+++ src/pages/personal/Personal.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Navigation from '../components/Navigation';
-import useLocalStorage from '../scripts/localStorageData';
+import Navigation from '../../components/navigation/Navigation';
+import useLocalStorage from '../../scripts/localStorageData';
+import "./Personal.css";
 
 function Personal() {             //Defining the function Personal in which all the constant defined which uses to store the values of User's Personal Information 
     const [valuePersonalFullname, setValuePersonalFullname] = useLocalStorage(
