+++ src/pages/privacy/Privacy.css
@@ -1,0 +1,22 @@
+#privacy-policy {
+    margin: 2rem auto;
+    width: 50%;
+}
+
+#privacy-policy a {
+    text-decoration: none;
+    color: hsl(219, 86%, 57%);
+}
+#privacy-policy a:hover {
+    text-decoration: underline;
+}
+
+#privacy-policy > h1, h2 {
+    margin-bottom: .8rem;
+    margin-top: .4rem;
+}
+
+#privacy-policy #footer {
+    position: initial;
+	margin-top: 2rem;
+}
