+++ src/pages/privacy/Privacy.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Footer from '../components/Footer';
+import Footer from '../../components/footer/Footer';
+import "./Privacy.css"
 
 function Privacy() {        //Funtion stated the Privacy Policy for the App
     return (
