+++ src/pages/programming/Programming.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { renderToString } from 'react-dom/server';
-import Navigation from '../components/Navigation';
-import useLocalStorage from '../scripts/localStorageData';
+import Navigation from '../../components/navigation/Navigation';
+import useLocalStorage from '../../scripts/localStorageData';
+import "./Programming.css";
 
 function Programming() {             //Defining the function Programming in which all the constants defined which takes the information about user's Prgramming skill
     const [valueProgramming, setValueProgramming] = useLocalStorage(
