+++ src/pages/result/Result.css
@@ -1,0 +1,4 @@
+#result .button-group {
+    display: flex;
+    align-items: center;
+}
