+++ src/pages/technologies/Technologies.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link, /* useLocation */ } from 'react-router-dom';
 import { renderToString } from 'react-dom/server';
-import Navigation from '../components/Navigation';
-import useLocalStorage from '../scripts/localStorageData';
+import Navigation from '../../components/navigation/Navigation';
+import useLocalStorage from '../../scripts/localStorageData';
+import "./Technologies.css";
 
 function Technologies() {           //This function defined the constants, used to store the value of technologies used by the user
     const [valueTechnologies, setValueTechnologies] = useLocalStorage(
